LIBSsa version: 2.0 alpha 23
MAJOR CHANGE: In this version I've decided to start to work in a new branch namely "updating_spectra_object", as I realized that some of the strictures were starting to become more and more complicated to deal as complexity/modules were added.
After some thought, I opted to use dictionaries to better store the elements in the Spectra object. Now, every major global variable, for example, "wavelengths" have two or more keys (in this case, 'Raw' and 'Isolated'). Spectra-wise information are in "samples" variable, containing the keys 'Count', 'Name' and 'path', and so on...
This will take some time to finish because all functions that operates in those objects must be rewritten. But in the end everything is going to be fine. ;-)

spectra.py
* Major update. Changed names of almost every global variable
* Added clear method
* Added prototype methods for load and save
* Changed objects from worker.py to here. Also, made some tweaks in the Worker and WorkerSignals objects (still testing)

worker.py
* Minor changes (will be removed eventually)

imports.py
* Updated load and outliers functions to properly work with the new structures

libssa.py
* Replaced Worker import from worker.py to spectra.py
* Updated functions until page 2 to work with the new Spectra object
** doplot (will need further updates)
** setgrange (will need further updates)
** spopen
** spload
** outliers
** loadrefcorrel (TODO: change name in the next commit)
** docorrel

libssa.ui
* Updated version: 2.0a23